requirements.txt
torch
torchvision
pytorch-ignite
pytorch-lightning
